Use ServerRunner w/ @main attribute vs main.swift
